Cleaned up some of the code, next steps is encoding arbitrary length data and working on the decoder plus possible key storage techniques.
Some thoughts on potential attack vectors: Potential side channel attack by sniffing data over time on simple layered versions of this encoding would allow for somewhat easy cracking. For example, it would be non trivial to build up a dictionary and reverse engineer the translation matrix on a single layer encoding scheme. However, this requires the attacker to be able to see both the encrypted and non encrypted channels and the threat is muted when adding more layers so it's just as non trivial to mitigate...